requirements.txt
transformers
python-dotenv
sqlalchemy
tika
pandas
PyPDF2
pillow
opencv-python
pytesseract
gradio

